Exit with FALSE handling 'help' argument.
'help' argument is a special case arg intended just to print help message and
exit the script. Therefore, you have to return FLAGS_FALSE and propagate this
to the user to give a chance to stop the script.

Typicall usage of shflags is like below

   FLAGS "$@" || exit $?

and it should do the job to exit script on '--help' as expected.

Signed-off-by: Waldemar Rymarkiewicz <waldemar.rymarkiewicz@gmail.com>